requirements.txt
flask
pandas
python-dotenv
openai
sarvamai